Fix early end infinite loop
`readLine()` returns `null` at the end of the file, but concatenating the
newline at the end converts it into the string `"null\n"`, thus prevent
`LexXML#IsEOF()` from ever returning true, since it just checks for a lexeme
of `null`. This lead to an infinite loop for XML documents that were not
terminated (the layout file John shared was cut short, so his debugger would
fail to start up!).

Fixes #509, though I plan to do more work on this.