Safer XML file writing
There are lots of chances for an IOException to occur while writing to a file
buffer, and I think the problem is made worse by the fact that the reporters
of #509 keep these files on OneDrive.

Write the entire XML document first to a temp file, then rename that file to
the correct final file name. This should help prevent us from borking the user's
layout file, since the worse that could happen is that the final rename fails,
preventing us from adding the user's new changes, and the user is warned of this
error.

Do the same process for the .ini config file, as well.